W1E10: Mark w1e10 as complete
The exercise stated that there should be only one click handler,
and that it should be implemented as HOF (Higher-order function).

However, this would cause React to create new function on every
render. Abetter way to solve the problem is by passing wanted
key to the child directly and have the child pass it to the
event handler.